requirement.txt
cmake
dilib
opencv-python
numpy
Pillow
matplotlib
face_recognition
scipy
playsound
streamlit